add Flower, Tree, Garden classes to garden project

+++ week-04/day-01/garden/Plant.cpp
@@ -3,7 +3,7 @@
 
 
 
-void Plant::water(float amount)
+void Plant::water(double amount)
 {
     _waterLevel += amount * 1;
 }
